Convert the buildbot.www.auth module into a package
