+++ master/buildbot/newsfragments/buildbot-master-www-auth-package.misc
@@ -1,0 +1,1 @@
+The :py:class:`~buildbot.www.auth` module was converted into a package.
