+++ master/buildbot/www/auth/__init__.py
@@ -1,0 +1,29 @@
+# This file is part of Buildbot.  Buildbot is free software: you can
+# redistribute it and/or modify it under the terms of the GNU General Public
+# License as published by the Free Software Foundation, version 2.
+#
+# This program is distributed in the hope that it will be useful, but WITHOUT
+# ANY WARRANTY; without even the implied warranty of MERCHANTABILITY or FITNESS
+# FOR A PARTICULAR PURPOSE.  See the GNU General Public License for more
+# details.
+#
+# You should have received a copy of the GNU General Public License along with
+# this program; if not, write to the Free Software Foundation, Inc., 51
+# Franklin Street, Fifth Floor, Boston, MA 02110-1301 USA.
+#
+# Copyright Buildbot Team Members
+
+
+from buildbot.www.auth.base import (
+    AuthRootResource, AuthBase, UserInfoProviderBase, LoginResource,
+    AuthRealm, PreAuthenticatedLoginResource, LogoutResource
+)
+from buildbot.www.auth.userpasswd import (
+    TwistedICredAuthBase, HTPasswdAuth, UserPasswordAuth
+)
+from buildbot.www.auth.customauth import CustomAuth
+from buildbot.www.auth.remoteuser import RemoteUserAuth
+
+
+class NoAuth(AuthBase):
+    pass
